==> Sinkhole-Intelligence-Capstone/03_Data_Wrangling/0_model_satellite_attr_consolidated.ipynb <==
import pandas as pd
import numpy as np
from zipfile import ZipFile
import os
from os.path import basename
import urllib
import shutil
from geopandas import gpd
from shapely.geometry import Point, Polygon
import datetime
from datetime import datetime, timedelta

# import shapefile

# my script
from w210_attribute_library import withinstates, haversine_distance, getdate, consolidate_sat_data

datdirsat = "../data/satellite_data/"
datdir = "../data/"
attrs = "../attrs/"
modeld = "../model/"

## Load Sinkhole Data - to get Date

df_sh = pd.read_excel(datdir+"Florida_Subsidence_Incident_Reports.xlsx", sheet_name=0)
df_sh["DateD"] = df_sh.apply(lambda row: getdate(row["EVENT_DATE"]), axis=1)
sh = df_sh[["OBJECTID", "DateD"]]
sh.columns = ["ID", "DateD"]
print(len(sh))
sh = sh.dropna()
print(len(sh))

### Process Satellite Data 365

nogood = ["2009_0_1", "428_0_1"]
dfsat365 = consolidate_sat_data(365, datdirsat,
                                "model_sh_attr_365_group_1_group_0.csv", 
                                "model_sh_attr_365_group_2_group_0.csv", sh, nogood)

dfsat365.to_csv(datdir+"model_satel_attr_365.csv", index=False)
pivot = np.round(pd.pivot_table(dfsat365, values='prediction', 
                                index='label', 
                                columns='Group', 
                                aggfunc='count'),2)
pivot

### Satellite Attributes - 60 days

nogood = ["2009_0_1", "428_0_1"]
dfsat60 = consolidate_sat_data(0, datdirsat,
                                "model_sh_attr_60_group_1_group_0.csv", 
                                "model_sh_attr_60_group_2_group_0.csv", sh, nogood)

dfsat60.to_csv(datdir+"model_satel_attr_60.csv", index=False)
pivot = np.round(pd.pivot_table(dfsat60, values='prediction', 
                                index='label', 
                                columns='Group', 
                                aggfunc='count'),2)
pivot

### Satellite Attributes - 0 days

nogood = ["2009_0_1", "428_0_1"]
dfsat0 = consolidate_sat_data(0, datdirsat,
                                "model_sh_attr_1_group_1_group_0.csv", 
                                "model_sh_attr_1_group_2_group_0.csv", sh, nogood)

dfsat0.to_csv(datdir+"model_satel_attr_0.csv", index=False)
pivot = np.round(pd.pivot_table(dfsat0, values='prediction', 
                                index='label', 
                                columns='Group', 
                                aggfunc='count'),2)
pivot

### OLD CODE FOR REFERENCE ONLY

dfsat60_g1_g0 = pd.read_csv(datdirsat+"model_sh_attr_60_group_1_group_0.csv")
dfsat60_g1_g0["Group"] =  np.where(dfsat60_g1_g0["label"]==1,0,1)

dfsat60_g2 = pd.read_csv(datdirsat+"model_sh_attr_60_group_2_group_0.csv")
dfsat60_g2 = dfsat60_g2[dfsat60_g2["label"] == 0]
dfsat60_g2["start_date"] = dfsat60_g2.apply(lambda r: r["start_date"][6:]+"-"+r["start_date"][0:2]+"-"+r["start_date"][3:5], axis=1)
dfsat60_g2["Group"] = 2
dfsat60_g2.head()

# dfsat60 = dfsat60_g1_g0.append(dfsat60_g2)
# dfsat60["DateD"] = dfsat60.apply(lambda row: datetime.strptime(row["start_date"],"%Y-%m-%d")+daysdelta, axis=1)

dfsat60["Key"] = dfsat60.apply(lambda row: str(row["ID"])+"_"+str(row["label"])+"_"+str(row["Group"]), axis=1)

nogood = ["2009_0_1", "428_0_1"]
dfsat60 = dfsat60[~dfsat60["Key"].isin(nogood)]

pivot = np.round(pd.pivot_table(dfsat60, values='prediction', 
                                index='label', 
                                columns='Group', 
                                aggfunc='count'),2)
pivot

dfsat60.to_csv(datdir+"model_satel_attr_60.csv", index=False)
