==> Sinkhole-Intelligence-Capstone/03_Data_Wrangling/1_w210_sinkhole_Attr_gen_Final.ipynb <==
# Generate Sinkhole Attributes

import numpy as np
import pandas as pd
import os
import datetime
from datetime import datetime, timedelta
import matplotlib.pyplot as plt
import math
import warnings
warnings.filterwarnings("ignore")

from w210_attribute_library import haversine_distance, getdate, findname, findAttr, shAttributes, sh_attr2

datdirsat = "../data/satellite_data/"
datdir = "../data/"
attrs = "../attrs/"
modeld = "../model/"

## Load Sinkhole Date for Florida (State)

df_sh = pd.read_excel(datdir+"Florida_Subsidence_Incident_Reports.xlsx", sheet_name=0)
df_sh["DateD"] = df_sh.apply(lambda row: getdate(row["EVENT_DATE"]), axis=1)
df_sh = df_sh[~df_sh["EVENT_DATE"].isna()]
print(len(df_sh))

## Load 365 Data

dfevent365 = pd.read_csv(datdir+"model_satel_attr_365.csv")
dfevent365["DateD"] = dfevent365.apply(lambda row: datetime.strptime(row["DateD"],"%Y-%m-%d"), axis=1)
dfevent365.head(1)

pivot = np.round(pd.pivot_table(dfevent365, values='prediction', 
                                index='label', 
                                columns='Group', 
                                aggfunc='count'),2)
pivot

## Generate Sinkhole Attributes (Using Tile Level Calculations)

daysdelta = timedelta(365)
fields = ["lon", "lat", "ID"]
sh365 = shAttributes(df_sh, dfevent365, daysdelta, fields)
sh365.to_csv(attrs+"model_sat_sh_attr_365.csv", index=False)

sh365.head(2)

### Sinkhole Attributes for 60 days

dfevent60 = pd.read_csv(datdir+"model_satel_attr_60.csv")
dfevent60["DateD"] = dfevent60.apply(lambda row: datetime.strptime(row["DateD"],"%Y-%m-%d"), axis=1)

## Generate Sinkhole Attributes
daysdelta = timedelta(60)
fields = ["lon", "lat", "ID"]
sh0 = shAttributes(df_sh, dfevent60, daysdelta, fields)
sh0.to_csv(attrs+"model_sat_sh_attr_60.csv", index=False)

pivot = np.round(pd.pivot_table(dfevent60, values='prediction', 
                                index='label', 
                                columns='Group', 
                                aggfunc='count'),2)
pivot

### Sinkhole Attributes for 0 days

dfevent0 = pd.read_csv(datdir+"model_satel_attr_0.csv")
dfevent0["DateD"] = dfevent0.apply(lambda row: datetime.strptime(row["DateD"],"%Y-%m-%d"), axis=1)

## Generate Sinkhole Attributes
daysdelta = timedelta(0)
fields = ["lon", "lat", "ID"]
sh0 = shAttributes(df_sh, dfevent0, daysdelta, fields)
sh0.to_csv(attrs+"model_sat_sh_attr_0.csv", index=False)
