==> Sinkhole-Intelligence-Capstone/03_Data_Wrangling/1_1_model_generate_Karst_attribute.ipynb <==
import pandas as pd
import numpy as np
from zipfile import ZipFile
import os
from os.path import basename
import urllib
import shutil
from geopandas import gpd
from shapely.geometry import Point, Polygon
# import shapefile

# my script
from w210_attribute_library import withinstates, haversine_distance

datdir = "../data/"
attrs = "../attrs/"
modeld = "../model/"

gdf = gpd.read_file("../data/sink_density_classified_polys_1km_conus/sink_density_classified_polys_1km_conus.shp")
# geo_df = GeoDataFrame(df, crs=crs, geometry=geometry)

new_gdf = gdf.to_crs(epsg=4326)
new_gdf.head(1)

new_gdf['x_coord'] = new_gdf.apply(lambda row: row['geometry'].centroid.x, axis=1)
new_gdf['y_coord'] = new_gdf.apply(lambda row: row['geometry'].centroid.y, axis = 1)

new_gdf.to_csv(datdir+"karst_gridcode_raw_data.csv", index=False)

### Read Karst Data
Note: if already created, load file here

fkarst = 'karst_gridcode_raw_data.csv'
dfk = pd.read_csv(datdir+fkarst)
print(len(dfk))
dfk.head(1)

### Select Relevant Geopoints

subdir = "../data/shapefile/"
shapedir = 'cb_2018_us_state_500k/'
shapefile500 = "cb_2018_us_state_500k.shp"

us500 = gpd.read_file(subdir+shapedir+shapefile500)

flgeometry = list(us500[(us500["NAME"]=='Florida')]["geometry"])[0]
gageometry = list(us500[(us500["NAME"]=='Georgia')]["geometry"])[0]
algeometry = list(us500[(us500["NAME"]=='Alabama')]["geometry"])[0]

geometries = [flgeometry, gageometry, algeometry ]
# geometries = [flgeometry]

# dfk["Florida"] = dfk.apply(lambda row: "FL" if (Point(row["x_coord"],row["y_coord"]).within(flgeometry)) else "NoFL", axis=1)

dfk["in_relevant_state"] = dfk.apply(lambda row: withinstates(geometries, Point(row["x_coord"],row["y_coord"])), axis=1)

dfk = dfk[dfk["in_relevant_state"] == "Yes"]
len(dfk)

### Read Tile Data

ftile = 'model_satel_attr_365.csv'
dtile = pd.read_csv(datdir+ftile)
dtile.columns

dtile = dtile[['Key', 'lon', 'lat']]

#### Cross Merge Dataframes

dtileK = pd.merge(dtile, dfk, how="cross")
print(len(dtileK))
dtileK.head(1)

#### Calculate Distance

dtileK['Distance'] = dtileK.apply(lambda row: 
                                    haversine_distance(row['lat'], row['lon'], 
                                                       row['y_coord'], row['x_coord'], 
                                                       earth_radius=3963.19), axis=1)
dtileK.head(2)

## Find the Minimum Distances within Tile and Weather Station

`df.groupby('Company')['MPG'].agg('min')`  
`df.groupby('Company')[['MPG', 'EngineSize']].agg('min')`

**Reference:**
https://datascienceparichay.com/article/pandas-groupby-minimum/

# Find the rows with the minimum Distance for each Key_x
dfmin1 = dtileK.groupby(['Key'])['Distance'].min().to_frame()
print(len(dfmin1))

# Select only the rows with the minimum
keysL = list(dfmin1.index)
minD = list(dfmin1['Distance'])
dfF1 = dtileK[((dtileK['Key'].isin(keysL)) &  (dtileK['Distance'].isin(minD)))]

#Checking for Duplicates
df2 = dfF1[dfF1["Key"].duplicated()==True]
dup1 = df2["Key"].unique()
# dup1 = ['1082_0_1','1083_0_1', '2406_0_1', '2459_0_1', '2463_0_1', '2737_0_1', '3294_0_2', '3294_1_0', '556_0_1']
dfF1[(dfF1['Key'].isin(dup1))]

dfF1.drop_duplicates(subset=['Key'], inplace=True)
dfF1["Distance"].describe()

dfF1["gridcode"].unique()

dfF1.columns

dfF1 = dfF1[['Key', 'gridcode']]

dfF1.to_excel(datdir+"w210_karst.xlsx", index=False)
