==> Sinkhole-Intelligence-Capstone/03_Data_Wrangling/2_w210_match_tile_to_ws.ipynb <==
# Haversine distance formula
import numpy as np
import pandas as pd
import datetime
from datetime import datetime, timedelta

from w210_attribute_library import haversine_distance, checkwsquality, ws_tile_pairs, ws_tile_pair_final

datdir = "../data/"
attrs = "../attrs/"

## Read Weather Data

fweather = 'weather_data.csv'
dfw = pd.read_csv(datdir+fweather)
dfw["DateD"] = dfw.apply(lambda row: datetime.strptime(str(row["year"])+"-"+str(row["mo"])+"-"+str(row["da"]),"%Y-%m-%d") ,axis=1)
print(len(dfw))

pivot = np.round(pd.pivot_table(dfw, values='prcp', 
                                index='year', 
                                columns='mo', 
                                aggfunc='count'),2)
pivot

wsnames = set(dfw["stn_wban"].unique())
ws = dfw[['stn_wban', "name", "lon", "lat"]].drop_duplicates()
ws.columns = ["Key", 'name', 'lon_w', 'lat_w']
print(len(wsnames),len(ws))

duplicateName = ['WHITING FIELD NAVAL AIR STATI', 'KEYSTONE AIRPARK']
ws[(ws['name'].isin(duplicateName))]

### Process 365 Data

df365 = ws_tile_pair_final(attrs+"model_sat_sh_attr_365.csv", ws, dfw)

# Final Quality Control Check
wsgood, wsnogood = checkwsquality(df365, dfw)
print(len(wsnogood))

# Save File - assuming wsnogood is ZERO
df365.to_csv(attrs+"model_wspair_sat_sh_attr_365.csv", index=False)

### Repeat Process for 60 days

df60 = ws_tile_pair_final(attrs+"model_sat_sh_attr_60.csv", ws, dfw)

# Final Quality Control Check
wsgood, wsnogood = checkwsquality(df60, dfw)
print(len(wsnogood))

# Save File - assuming wsnogood is ZERO
df60.to_csv(attrs+"model_wspair_sat_sh_attr_60.csv", index=False)

## Repeat Process for 0 days

df0 = ws_tile_pair_final(attrs+"model_sat_sh_attr_0.csv", ws, dfw)

# Final Quality Control Check
wsgood, wsnogood = checkwsquality(df0, dfw)
print(len(wsnogood))

# Save File - assuming wsnogood is ZERO
df0.to_csv(attrs+"model_wspair_sat_sh_attr_0.csv", index=False)
