==> Sinkhole-Intelligence-Capstone/03_Data_Wrangling/3_w210_soil_attr_to_tile.ipynb <==
import numpy as np
import pandas as pd
import os

from w210_attribute_library import haversine_distance

datadir = '../data/'
attrs = "../attrs/"

def soil_attr(fsoil, datadir, attrs, ftile):
    
    soild = pd.read_csv(datadir+fsoil)

    # Read Feature Table
    dtil = pd.read_csv(attrs+ftile)
 
    # Cross Merge Dataframes
    dtil = pd.merge(dtil,soild, how="cross")

    # Get Distance
    dtil['Distance'] = dtil.apply(lambda row: 
                                        haversine_distance(row['lat'], row['lon'], 
                                                           row['Center Latitude'], 
                                                           row['Center Longitude'], 
                                                           earth_radius=3963.19), axis=1)
    
    dfmin1 = dtil.groupby(['Key_x'])['Distance'].min().to_frame()

    keysL = list(dfmin1.index)
    minD = list(dfmin1['Distance'])
    dtil = dtil[((dtil['Key_x'].isin(keysL)) &  (dtil['Distance'].isin(minD)))]
    
    return dtil

## Process 365 delta

df365 = soil_attr('FL_soil_byCounty.csv', datadir, attrs, "model_wspair_sat_sh_attr_365.csv")
df365.to_csv(attrs+"model_wspair_sat_sh_soil_attr_365.csv", index=False)

## Process 60 delta

df60 = soil_attr('FL_soil_byCounty.csv', datadir, attrs, "model_wspair_sat_sh_attr_60.csv")
df60.to_csv(attrs+"model_wspair_sat_sh_soil_attr_60.csv", index=False)

## Process 0 delta

df0 = soil_attr('FL_soil_byCounty.csv', datadir, attrs, "model_wspair_sat_sh_attr_0.csv")
df0.to_csv(attrs+"model_wspair_sat_sh_soil_attr_0.csv", index=False)
