==> Sinkhole-Intelligence-Capstone/03_Data_Wrangling/4_w210_get_ws_attributes_model.ipynb <==
## Get Weather Attributes from Identified Weather Station

# Haversine distance formula
import numpy as np
import pandas as pd
import datetime
from datetime import datetime, timedelta

from w210_attribute_library import getwf_final, weatherFeatures

## Read Weather Data

datadir = "../data/"
attrs = "../attrs/"
fweather = 'weather_data.csv'
dfw = pd.read_csv(datadir+fweather)
print(len(dfw))

len(list(dfw["stn_wban"].unique()))

pivot = np.round(pd.pivot_table(dfw, values='prcp', 
                                index='year', 
                                columns='mo', 
                                aggfunc='count'),2)
pivot

colweather = ['stn_wban', 'year', 'mo', 'da', 'temp', 'prcp', 'flag_prcp', 'min', 'max',
       'flag_prcp', 'rain_drizzle', 'sndp', 'fog', 'snow_ice_pellets', 'hail',
       'thunder', 'tornado_funnel_cloud', 'wdsp', 'mxpsd', 'gust', 'visib',
       'dewp', 'slp', 'usaf_wban', 'wban']
dfw1 = dfw[colweather]
dfw1.head(1)

# Create a Date Field to be able to make date comparisons
dfw1["DateD"] = dfw1.apply(lambda row: datetime.strptime(str(row["year"])+"-"+str(row["mo"])+"-"+str(row["da"]),"%Y-%m-%d") ,axis=1)

dfw1["DateD"].min()

dfw1.describe()

# Replace values with NaN
no_valid = [99.99000, 999.900000, 9999.900000, 99999.000000]
dfw1 = dfw1.replace(99.99000,np.nan)
dfw1 = dfw1.replace(999.9000000,np.nan)
dfw1 = dfw1.replace(9999.900000,np.nan)
dfw1 = dfw1.replace(99999.000000,np.nan)

dfw1.describe()

dfw1.head()

## Read Tile-Weather Station Pair Data

df365 = pd.read_csv(attrs+ "model_wspair_sat_sh_soil_attr_365.csv")
df365["DateD"] = df365.apply(lambda row: datetime.strptime(row["DateD"],"%Y-%m-%d"), axis=1)

## Get Weather Features

### Get Weather Attributes with timedelta = 365

# date_string = '2022/07/05'

date_string = 'none'
daysdelta = timedelta(365)
is_scale = False

dfwf = weatherFeatures(df365, dfw1, is_scale, daysdelta, date_string)

dfwf.to_csv(attrs + "model_ws_attr_365.csv", index=False)

**Reference:** https://www.geeksforgeeks.org/different-ways-to-create-pandas-dataframe/

### Get Weather Feature with timedelta = 0

# date_string = '2022/07/05'

df60 = pd.read_csv(attrs+ "model_wspair_sat_sh_soil_attr_60.csv")
df60["DateD"] = df60.apply(lambda row: datetime.strptime(row["DateD"],"%Y-%m-%d"), axis=1)
df60.head(1)


date_string = 'none'
daysdelta = timedelta(0)
is_scale = False

dfwf = weatherFeatures(df60, dfw1, is_scale, daysdelta, date_string)

dfwf.head(1)

dfwf.to_csv(attrs + "model_ws_attr_60.csv", index=False)

### Process 0 days delta

# date_string = '2022/07/05'

df0 = pd.read_csv(attrs+ "model_wspair_sat_sh_soil_attr_0.csv")
df0["DateD"] = df0.apply(lambda row: datetime.strptime(row["DateD"],"%Y-%m-%d"), axis=1)
df0.head(1)


date_string = 'none'
daysdelta = timedelta(0)
is_scale = False

dfwf = weatherFeatures(df0, dfw1, is_scale, daysdelta, date_string)

dfwf.to_csv(attrs + "model_ws_attr_0.csv", index=False)

# NOT USED ANYMORE - Join Rolling Features to Main Model Input

finEvents = pd.read_csv(datadir+"w210_model_input_sh_sat_wf_kars_soil_F.csv")
print(len(finEvents))

finEvents.head(2)

finEvents = pd.merge(finEvents, dfwf, on="Key")
len(finEvents)

finEvents.head(2)

finEvents.to_csv(datadir + "w210_model_input_sh_sat_wfRolling_kars_soil_F.csv", index=False)
