==> Sinkhole-Intelligence-Capstone/03_Data_Wrangling/5_w210_join_weather_and_karst_attrs.ipynb <==
import numpy as np
import pandas as pd

datadir = "../data/"
attrs = "../attrs/"
model = "../model/"

## Read Current Sinkhole/Soil 365 delta and Satellite Attribute File

minput1 = pd.read_csv(attrs+"model_wspair_sat_sh_soil_attr_365.csv")

# Read File SH and Soil attribute with 0 days delta
minput1 = pd.read_csv(attrs+"model_wspair_sat_sh_soil_attr_365.csv")

# Rename columns
minput1 = minput1.rename(columns={'Key_x': 'Key', 'lon': 'lon_t', 'lat': 'lat_t'})

# Read File WS feature 9 days delta
wsf = pd.read_csv(attrs+"model_ws_attr_365.csv")

# Merge Files
finEvents = pd.merge(minput1, wsf, on="Key")

# Read Karst Indicator file
karst = pd.read_excel(datadir+"w210_karst.xlsx")

# Merge Karst File
finEvents = pd.merge(finEvents, karst, on="Key")


finEvents = finEvents[['name_x', 'DateD', 'Key', 'imgnum', 'label', 'Group', 'ID', 
       'lon_t_x', 'lat_t_x', 'geometry', 'AnnualCrop', 'Forest', 'HerbaceousVegetation', 'Highway',
       'Industrial', 'Pasture', 'PermanentCrop', 'Residential', 'River',
       'SeaLake', 'prediction', 'prediction_name', 
       'l25', 'l50', 'l75', 'l100', 'l150', 'l200', 'l250', 'l300', 'l500',
       'l750', 'l1000', 'l1000plus', 'coloc', 'Y25', 'Y50', 'Y75', 'Y100',
       'Y150', 'Y200', 'Y250', 'Y300', 'Y500', 'Y750', 'Y1000', 'Y1000plus',
       'Ycoloc', 'Key_y_x', 'name_y', 'lon_w_x', 'lat_w_x', 
       'County', 'Calcium Carbonate', 'Gypsum', 'Soil Health Organic Matter', 
       'Percent Clay', 'Percent Sand', 'Percent Silt', 'Available Water Storage',
       'rolling_7_precip', 'rolling_15_precip', 'rolling_30_precip',
       'rolling_60_precip', 'rolling_90_precip', 'y1_mean_prc', 'y1_max_prc',
       'y1_mean_tmp', 'y1_max_tmp', 'y1_min_tmp', 'y2_mean_prc', 'y2_max_prc',
       'y2_mean_tmp', 'y2_max_tmp', 'y2_min_tmp', 'y3_mean_prc', 'y3_max_prc',
       'y3_mean_tmp', 'y3_max_tmp', 'y3_min_tmp', 'gridcode']]

finEvents = finEvents.rename(columns={'label': 'Sinkhole','lon_t_x': 'lon_t', 'lat': 'lat_t_x',
                                  'Key_y_x': 'Key_ws', 'name_y':'ws_name', 
                                  'lon_w_x':'lon_w', 'lat_w_x':'lat_w'})

finEvents.to_csv(attrs+"model_wspair_sat_sh_soil_ws_karst_attr_365.csv", index=False)
finEvents.to_csv(model+"data_model1_365_unbalanced.csv", index=False)

## Read Current Sinkhole/Soil 60 days delta 

# Read File SH and Soil attribute with 0 days delta
minput1 = pd.read_csv(attrs+"model_wspair_sat_sh_soil_attr_60.csv")

# Rename columns
minput1 = minput1.rename(columns={'Key_x': 'Key', 'lon': 'lon_t', 'lat': 'lat_t'})

# Read File WS feature 9 days delta
wsf = pd.read_csv(attrs+"model_ws_attr_60.csv")

# Merge Files
finEvents = pd.merge(minput1, wsf, on="Key")

# Read Karst Indicator file
karst = pd.read_excel(datadir+"w210_karst.xlsx")

# Merge Karst File
finEvents = pd.merge(finEvents, karst, on="Key")


finEvents = finEvents[['name_x', 'DateD', 'Key', 'imgnum', 'label', 'Group', 'ID', 
       'lon_t_x', 'lat_t_x', 'geometry', 'AnnualCrop', 'Forest', 'HerbaceousVegetation', 'Highway',
       'Industrial', 'Pasture', 'PermanentCrop', 'Residential', 'River',
       'SeaLake', 'prediction', 'prediction_name', 
       'l25', 'l50', 'l75', 'l100', 'l150', 'l200', 'l250', 'l300', 'l500',
       'l750', 'l1000', 'l1000plus', 'coloc', 'Y25', 'Y50', 'Y75', 'Y100',
       'Y150', 'Y200', 'Y250', 'Y300', 'Y500', 'Y750', 'Y1000', 'Y1000plus',
       'Ycoloc', 'Key_y_x', 'name_y', 'lon_w_x', 'lat_w_x', 
       'County', 'Calcium Carbonate', 'Gypsum', 'Soil Health Organic Matter', 
       'Percent Clay', 'Percent Sand', 'Percent Silt', 'Available Water Storage',
       'rolling_7_precip', 'rolling_15_precip', 'rolling_30_precip',
       'rolling_60_precip', 'rolling_90_precip', 'y1_mean_prc', 'y1_max_prc',
       'y1_mean_tmp', 'y1_max_tmp', 'y1_min_tmp', 'y2_mean_prc', 'y2_max_prc',
       'y2_mean_tmp', 'y2_max_tmp', 'y2_min_tmp', 'y3_mean_prc', 'y3_max_prc',
       'y3_mean_tmp', 'y3_max_tmp', 'y3_min_tmp', 'gridcode']]

finEvents = finEvents.rename(columns={'label': 'Sinkhole','lon_t_x': 'lon_t', 'lat': 'lat_t_x',
                                  'Key_y_x': 'Key_ws', 'name_y':'ws_name', 
                                  'lon_w_x':'lon_w', 'lat_w_x':'lat_w'})

finEvents.to_csv(attrs+"model_wspair_sat_sh_soil_ws_karst_attr_60.csv", index=False)

finEvents.to_csv(model+"data_model1_60_unbalanced.csv", index=False)

### Process 0 Delta

# Read File SH and Soil attribute with 0 days delta
minput1 = pd.read_csv(attrs+"model_wspair_sat_sh_soil_attr_0.csv")

# Rename columns
minput1 = minput1.rename(columns={'Key_x': 'Key', 'lon': 'lon_t', 'lat': 'lat_t'})

# Read File WS feature 9 days delta
wsf = pd.read_csv(attrs+"model_ws_attr_0.csv")

# Merge Files
finEvents = pd.merge(minput1, wsf, on="Key")

# Read Karst Indicator file
karst = pd.read_excel(datadir+"w210_karst.xlsx")

# Merge Karst File
finEvents = pd.merge(finEvents, karst, on="Key")


finEvents = finEvents[['name_x', 'DateD', 'Key', 'imgnum', 'label', 'Group', 'ID', 
       'lon_t_x', 'lat_t_x', 'geometry', 'AnnualCrop', 'Forest', 'HerbaceousVegetation', 'Highway',
       'Industrial', 'Pasture', 'PermanentCrop', 'Residential', 'River',
       'SeaLake', 'prediction', 'prediction_name', 
       'l25', 'l50', 'l75', 'l100', 'l150', 'l200', 'l250', 'l300', 'l500',
       'l750', 'l1000', 'l1000plus', 'coloc', 'Y25', 'Y50', 'Y75', 'Y100',
       'Y150', 'Y200', 'Y250', 'Y300', 'Y500', 'Y750', 'Y1000', 'Y1000plus',
       'Ycoloc', 'Key_y_x', 'name_y', 'lon_w_x', 'lat_w_x', 
       'County', 'Calcium Carbonate', 'Gypsum', 'Soil Health Organic Matter', 
       'Percent Clay', 'Percent Sand', 'Percent Silt', 'Available Water Storage',
       'rolling_7_precip', 'rolling_15_precip', 'rolling_30_precip',
       'rolling_60_precip', 'rolling_90_precip', 'y1_mean_prc', 'y1_max_prc',
       'y1_mean_tmp', 'y1_max_tmp', 'y1_min_tmp', 'y2_mean_prc', 'y2_max_prc',
       'y2_mean_tmp', 'y2_max_tmp', 'y2_min_tmp', 'y3_mean_prc', 'y3_max_prc',
       'y3_mean_tmp', 'y3_max_tmp', 'y3_min_tmp', 'gridcode']]

finEvents = finEvents.rename(columns={'label': 'Sinkhole','lon_t_x': 'lon_t', 'lat': 'lat_t_x',
                                  'Key_y_x': 'Key_ws', 'name_y':'ws_name', 
                                  'lon_w_x':'lon_w', 'lat_w_x':'lat_w'})

finEvents.to_csv(attrs+"model_wspair_sat_sh_soil_ws_karst_attr_0.csv", index=False)

finEvents.to_csv(model+"data_model1_0_unbalanced.csv", index=False)
